# Reduce syscalls for performance